requirements.txt
flask
flask_sqlalchemy
flask_login
werkzeug
pillow
tensorflow
numpy

